refactor: checkAvailableUserForDelete 메소드 삭제

+++ src/main/java/techcourse/myblog/service/CommentService.java
@@ -26,7 +26,7 @@ public class CommentService {
 
     public void deleteComment(Long commentId, User user) {
         Comment comment = commentRepository.findById(commentId).orElseThrow(NotFoundObjectException::new);
-        comment.checkAvailableUserForDelete(user);
+        comment.checkCorrespondingAuthor(user);
         commentRepository.deleteById(commentId);
     }
